Fix permission issues + PR comments
